Set switch interval in a loop with increasing numbers
After talking to Pablo and Thomas about this, they suggested that
we start at a 1ns switch interval and go up to 100ns for every time
setting it fails. After three failures, we go back to original
switch interval.

This makes sure that the tests will run in parallel, even if the
switch interval is not changed.